requirements.txt
asyncpg==0.29.0                     # PostgreSQL adapter for asynchronous operations
bcrypt==4.1.3                       # password hashing
fastapi[standard]==0.115.6
pydantic-settings==2.3.4
pydantic==2.7.4
pyjwt[crypto]==2.8.0                # generate and verify the JWT tokens, imported as `jwt`
python-dotenv==1.0.1                # load environment variables from .env file
requests==2.32.3
SQLAlchemy[asyncio]==2.0.32         # [asyncio] allows greenlet to be installed on Apple M1 devices.